refactor(cli): extract shared metrics rendering and fix acl percentiles
The acl and fwstate CLIs each carried a copy pasted metrics layer. Move the
domain types, the formatters and the shared gRPC sections into the CLI core
library, leaving each CLI only its module specific counter grouping.

The two copies had silently diverged. Shared histogram values carry raw per
bucket counts rather than cumulative ones, a deliberate divergence from
Prometheus semantics documented at the producer. The fwstate copy accumulated
them correctly while the acl copy did not, so the gRPC latency percentiles the
acl CLI reported were wrong. Unifying on the correct implementation fixes them.
