+++ cli/core/Cargo.toml
@@ -20,6 +20,7 @@ prost-types = "0.13"
 tonic = { version = "0.13", features = ["gzip"] }
 ynpb = { path = "../../common/rust/ynpb", version = "0.1", package = "yanet-ynpb" }
 readinesspb = { path = "../../common/rust/readinesspb", version = "0.1", package = "yanet-readinesspb" }
+commonpb = { path = "../../common/rust/commonpb", version = "0.1", package = "yanet-commonpb" }
 tower = "0.5"
 http = "1"
 ssh-key = "0.6"
