+++ cli/core/src/lib.rs
@@ -8,6 +8,7 @@ pub mod display;
 pub mod errors;
 pub mod humanfmt;
 pub mod logging;
+pub mod metrics;
 pub mod output;
 pub mod readiness;
